+++ cmd/install_darwin_arm64.sh
@@ -27,8 +27,18 @@ PLATFORM="darwin_arm64"
 
 # Base installation directory:
 GOPJRT_INSTALL_DIR="${GOPJRT_INSTALL_DIR:-/usr/local}"
+echo "Installing GoPJRT C/C++ dependencies under ${GOPJRT_INSTALL_DIR} (lib/ and include/ sub-directories)"
+echo "  - Set GOPJRT_INSTALL_DIR if you want another directory."
+
+# Should it use 'sudo' while installing ?
 _SUDO="sudo"
 if [[ "${GOPJRT_NOSUDO}" != "" ]] ; then
+  echo "  - Not using sudo during installation, disabled with GOPJRT_NOSUDO != ''."
+  _SUDO=""
+elif command -v sudo ; then
+  echo "  - Using sudo when extracting files to final destination (Set GOPJRT_NOSUDO=1 if you don't want sudo to be used)"
+else
+  echo "  - Not using sudo during installation, no program 'sudo' found in PATH."
   _SUDO=""
 fi
 
@@ -53,7 +63,7 @@ if [[ "${_SUDO}" != "" ]] ; then
   echo "Checking sudo authorization for installation"
   ${_SUDO} printf "\tsudo authorized\n"
 fi
-sudo tar xvzf "${tar_file}"
+${_SUDO} tar xvzf "${tar_file}"
 rm -f "${tar_file}"
 
 popd